refactor: Validator 가 가졌던 책임을 검증 대상 객체가 가지게 함.

+++ src/main/java/com/programmers/voucher/domain/voucher/model/Voucher.java
@@ -6,5 +6,5 @@ public interface Voucher {
 
 	UUID getVoucherId();
 
-	int discount(int beforeDiscount);
+	double discount(double beforeDiscount);
 }
